Ensure AGC algorithm always has an initial value
If np.abs(sample) != 0 for the first sample, the else branch
is hit, resulting in a NameError since the level_db variable is not
yet defined. This fix also sets an initial value for "level_db"
because static checkers can't generally figure out that the else
branch can no longer be hit on the first loop iteration.